+++ noodles-cram/CHANGELOG.md
@@ -8,11 +8,15 @@
 
 ### Changed
 
+  * cram/reader/container/block: Treat blocks with no data as empty, regardless
+    of compression method ([#131]).
+
   * cram/reader/header_container: Require blocks to be either uncompressed or
     gzipped.
 
     This was clarified in [samtools/hts-specs@4b61c4d].
 
+[#131]: https://github.com/zaeleus/noodles/issues/131
 [samtools/hts-specs@4b61c4d]: https://github.com/samtools/hts-specs/commit/4b61c4d7e40748b274478fd481bbfe60592a969b
 
 ## 0.21.0 - 2022-10-28